WpSource: Move `transformLink` to its own file
